修改运行问题

+++ core/data/dataloader/qgamedata.py
@@ -37,7 +37,7 @@ class LiveSegmentation(SegmentationDataset):
         # general resize, normalize and toTensor
         # if self.transform is not None:
         #     img = self.transform(img)
-        img = img / 127.5 - 1
+        img = np.float32(img) / 127.5 - 1
         input_tensor = transforms.ToTensor()
         img = input_tensor(img)
         mask = mask[:,:,0]
